requirements.txt
Python
Django
dgango-markdownx
django-taggit==1.1.0